Split test_pauli.py and test_hamiltonian.py out of test_operator.py
